Split out nested classes from `TimelineBlueprintContainer`
They got too big.
